Improve animations, add walking sound and fix a few bugs
For the animations, I have slightly modified the design of the player,
and added some bouncing effects while he walks

I have also removed pixel snapping as it was pretty buggy and is a lot
less smooth than the alternative. I will keep it for any still objects,
however.

+++ src/Tile.zig
@@ -3,6 +3,6 @@ const rl = @import("raylib");
 pub const Tile = struct {
     pub const size = 12;
 
-    hardness: i32 = 0,
+    hardness: u8 = 0,
     texture: *rl.Texture = undefined,
 };
